# Conflicts:
#	ARCHITECTURE.md
#	docs/architecture/DATA_FLOW_PATHS.md
#	docs/architecture/MODULE_ARCHITECTURE.md
#	docs/architecture/MODULE_INTEGRATION_CONTRACTS.md
#	docs/architecture/RELEASE_ARCHITECTURE_STATUS.md
#	src/CROSS_MODULE_INTEGRATION.md

+++ src/analytics/model_serving.cpp
@@ -127,7 +127,7 @@ void recordLatency(ModelServingEntry &e, double ms, size_t window) {
     }
 
     e.latency_buf.push_back(ms);
-    if (e.latency_buf.size() > window) {
+    if (static_cast<int>(e.latency_buf.size()) > window) {
         e.latency_buf.pop_front();
     }
 
@@ -222,7 +222,7 @@ void ModelServingEngine::registerModel(const std::string &name, const std::strin
 
     std::unique_lock lock(impl_->mu);
 
-    if (impl_->registry.size() >= static_cast<size_t>(impl_->config.max_models)) {
+    if (impl_->registry.size() >= impl_->config.max_models) {
         throw std::runtime_error(
             "ModelServingEngine: registry is full (max_models=" + std::to_string(impl_->config.max_models) + ")");
     }
@@ -297,7 +297,7 @@ std::string ModelServingEngine::predict(const std::string &name, const std::stri
 
 std::vector<std::string> ModelServingEngine::predictBatch(const std::string &name, const std::string &version,
                                                           const std::vector<DataPoint> &data) const {
-    if (data.size() > impl_->config.max_batch_size) {
+    if (static_cast<int>(data.size()) > impl_->config.max_batch_size) {
         throw std::invalid_argument("ModelServingEngine: batch size " + std::to_string(data.size())
                                     + " exceeds max_batch_size=" + std::to_string(impl_->config.max_batch_size));
     }
@@ -330,7 +330,7 @@ std::vector<std::string> ModelServingEngine::predictBatch(const std::string &nam
 std::vector<std::map<std::string, double>> ModelServingEngine::predictProba(const std::string &name,
                                                                             const std::string &version,
                                                                             const std::vector<DataPoint> &data) const {
-    if (data.size() > impl_->config.max_batch_size) {
+    if (static_cast<int>(data.size()) > impl_->config.max_batch_size) {
         throw std::invalid_argument("ModelServingEngine: batch size " + std::to_string(data.size())
                                     + " exceeds max_batch_size=" + std::to_string(impl_->config.max_batch_size));
     }
